requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
PyJWT==2.8.0
tweepy==4.14.0
APScheduler==3.10.4
boto3==1.34.34
gunicorn==21.2.0
Werkzeug==3.0.1
SQLAlchemy==2.0.25
pytest==7.4.4  # For testing